feat(sidecar)!: stream invoke/raw bodies and add route-to-Forst middleware
Add createRouteToForstMiddleware with route-request-stream JSON (prefix + streamed
base64 body + suffix), invokeFunctionRawWithReadableBody, and split ForstSidecar
into forst-sidecar; document and test the flow.

BREAKING CHANGE: pipeInvokeRawStream(packageName, functionName, body, destination) now
takes body: ReadableStream<Uint8Array> instead of args: unknown[]. Build the stream
from JSON.stringify([args]) or from createInvokeRawJsonBodyReadable for Express.

Example:

```typescript
    import { Readable } from "node:stream";
    const body = Readable.toWeb(nodeReadable) as ReadableStream<Uint8Array>;
    await client.pipeInvokeRawStream("main", "StreamFn", body, res);
```

docs: rewrite root README for sidecar-first positioning

Reframe intro, comparisons, migration walkthrough, and package pointers around Node +
sidecar adoption.

refactor(devserver): extract dev HTTP server into internal/devserver

Move the dev server out of forst/cmd/forst into forst/internal/devserver: dedicated
handlers (health, version, functions, invoke, invoke/raw), shared response helpers,
invoke execution, and tests. Drop the old dev_server_http* surface from cmd in favor
of the internal package and slimmer runtime wiring.
