+++ forst/internal/devserver/response.go
@@ -1,0 +1,34 @@
+package devserver
+
+import (
+	"encoding/json"
+	"fmt"
+	"net/http"
+)
+
+// sendJSONResponse sends a JSON response to the client.
+func (s *DevServer) sendJSONResponse(w http.ResponseWriter, response DevServerResponse) {
+	w.Header().Set("Content-Type", "application/json")
+
+	if s.httpOpts.CORS {
+		w.Header().Set("Access-Control-Allow-Origin", "*")
+		w.Header().Set("Access-Control-Allow-Methods", "GET, POST, OPTIONS")
+		w.Header().Set("Access-Control-Allow-Headers", "Content-Type")
+	}
+
+	encoder := json.NewEncoder(w)
+	if err := encoder.Encode(response); err != nil {
+		http.Error(w, fmt.Sprintf("Failed to encode response: %v", err), http.StatusInternalServerError)
+	}
+}
+
+// sendError sends an error response to the client.
+func (s *DevServer) sendError(w http.ResponseWriter, errorMsg string, statusCode int) {
+	response := DevServerResponse{
+		Success: false,
+		Error:   errorMsg,
+	}
+
+	w.WriteHeader(statusCode)
+	s.sendJSONResponse(w, response)
+}
